Added Search.r script
This TBOG - Pulse - Search.r script is able to replace the individual scripts placed in each folder from its place in the recording sets folder (eg. "Outlast" which contains "Outlast - Part 01," "Outlast - Part 02," etc.), import the data, process the individual recordings if the Timed.csv and Hist.png files need to be created, and process them all together. This is a new feature this script offers; the ability to create a single, faceted graph to show the histograms of each run. Though I have not done so, it can also be used to produce summary statistics for each recording into a single file.